TODO: 添加数据库有关消息的 SQL 以及一些基本函数 （如 添加消息，修改消息，删除消息）

+++ BuddhismWebsite/src/java/com/buddhism/dao/postDaoImpl.java
@@ -152,7 +152,7 @@ public class postDaoImpl extends HibernateDaoSupport implements postDao
        
        Query query = s.createQuery("update Post p set p.postUp = :up where p.id = :id");
        query.setParameter("up", update);
-       query.setParameter("id", (short)id);
+       query.setParameter("id", id);
        
        query.executeUpdate();
